[ISSofaGUI] Add EnumFlagsDataWidget (//Ref// T9937)
This widget is responsible for displaying a grid of all possible flags
of an enum. It detects the content of the enum through the
AbstractTypeInfo API.

You can opt into using it by setting the widget of a BaseData to
"enum_flags".

(cherry picked from commit 12b397b517d8f68aa6f1a8bb114c94682ebfba31)